TRN /refactor : TopAppBar padding
